feat: скриншотер отделён в свой пакет
