+++ packages/core/.storybook/preview.ts
@@ -4,8 +4,6 @@ import React from 'react';
 
 import '../global.css';
 import './global-storybook.css';
-import { loadComments } from './addon/common/comments';
-import { withScreenshotsDecorator } from './addon/screenshots-comparator/decorator';
 
 import { withReactQueryDecorator } from '../components/react-query-provider/mocked';
 import { withMockedApi } from '../hooks/use-api/mocked';
@@ -20,7 +18,6 @@ export default {
     withMockedApi(),
     withMockedRouter(),
     withMockedSchemaLoader(),
-    withScreenshotsDecorator(),
     (Story) => {
       message.destroy();
       return React.createElement(Story);
@@ -33,34 +30,6 @@ export default {
         color: /(background|color)$/i,
         date: /Date$/i,
       },
-    },
-  },
-
-  loaders: [
-    async () => {
-      // @ts-expect-error -- всё хорошо, это нужно чтобы в тесте потом обратиться к этому значению
-      if (window.pullRequestComments) {
-        // @ts-expect-error -- всё хорошо, это нужно чтобы в тесте потом обратиться к этому значению
-        return { comments: window.pullRequestComments };
-      }
-
-      const pullRequestNumber = Number(process.env.PULL_REQUEST_NUMBER);
-
-      if (!pullRequestNumber || isNaN(pullRequestNumber)) {
-        return { comments: [] };
-      }
-
-      const comments = await loadComments(pullRequestNumber);
-      // @ts-expect-error -- всё хорошо, это нужно чтобы в тесте потом обратиться к этому значению
-      window.pullRequestComments = comments;
-
-      return { comments };
-    },
-  ],
-
-  globalTypes: {
-    screenshotsComparatorActive: {
-      defaultValue: false,
     },
   },
 } satisfies Preview;
