+++ packages/core/.storybook/test-runner.ts
@@ -1,3 +1,3 @@
-import { getScreenshoterConfig } from './addon/test-runner-config';
+import { getScreenshoterConfig } from '../../sauron-eye-addon/addon/test-runner-config';
 
 export default getScreenshoterConfig();
