+++ packages/core/package.json
@@ -5,8 +5,7 @@
   "scripts": {
     "storybook": "storybook dev -p 6006",
     "storybook:build": "storybook build",
-    "test:interactions": "test-storybook --json --outputFile .storybook/screenshots/report.json",
-    "test:interactions:local": "wait-storybook --url 'http://localhost:6006/' && npm run test:interactions",
+    "test:interactions": "test-storybook --json --outputFile ./screenshots/report.json",
     "storybook:serve": "http-server ./storybook-static --port 6006 --silent"
   },
   "dependencies": {
@@ -17,24 +16,20 @@
     "@colibrijs/types": "^1.0.0",
     "@module-federation/utilities": "^3.0.13",
     "@tanstack/react-query": "^5.45.1",
-    "antd": "^5.18.3",
-    "classnames": "^2.5.1",
-    "resemblejs": "^5.0.0"
+    "antd": "^5.18.3"
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
     "@colibrijs/mocks": "^1.0.0",
+    "@colibrijs/sauron-eye-addon": "^1.0.0",
     "@colibrijs/wait-storybook": "^1.0.0",
     "@storybook/addon-actions": "^8.1.10",
     "@storybook/addon-essentials": "^8.1.10",
     "@storybook/addon-interactions": "^8.1.10",
     "@storybook/addon-links": "^8.1.10",
     "@storybook/addon-styling": "^1.3.7",
-    "@storybook/icons": "^1.2.12",
     "@storybook/addon-webpack5-compiler-swc": "^1.0.4",
     "@storybook/blocks": "^8.1.10",
-    "@storybook/components": "^8.1.0",
-    "@storybook/manager-api": "^8.1.0",
     "@storybook/react": "^8.1.10",
     "@storybook/react-webpack5": "^8.1.10",
     "@storybook/test": "^8.1.10",
