+++ packages/sauron-eye-addon/addon/addon.tsx
