+++ packages/sauron-eye-addon/addon/common/get-approved-screenshots.ts
