+++ packages/sauron-eye-addon/addon/common/get-report.ts
