+++ packages/sauron-eye-addon/addon/compare-button/compare-button.tsx
