+++ packages/sauron-eye-addon/addon/screenshot-panel/screenshot-panel.css
