+++ packages/sauron-eye-addon/addon/screenshot-panel/screenshot-table/screenshot-table.tsx
