+++ packages/sauron-eye-addon/addon/screenshots-comparator/decorator.tsx
