+++ packages/sauron-eye-addon/addon/screenshots-comparator/screenshots-comparator.tsx
