+++ packages/sauron-eye-addon/addon/test-runner-config/index.ts
@@ -1,1 +1,3 @@
 export { getScreenshoterConfig } from './screenshoter-config';
+
+export { isDirectoryAvailable } from './fs-utils';
