+++ packages/sauron-eye-addon/manager.ts
