+++ packages/sauron-eye-addon/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@colibrijs/sauron-eye-addon",
+  "version": "1.0.0",
+  "exports": {
+    "./preset": "./.storybook/preset.ts",
+    "./manager": "./.storybook/manager.ts",
+    "./package.json": "./package.json",
+    "./preview": "./.storybook/preview.ts",
+    "./test-runner-config": "./addon/test-runner-config/index.ts"
+  },
+  "dependencies": {
+    "@storybook/components": "^8.1.0",
+    "@storybook/icons": "^1.2.12",
+    "@storybook/manager-api": "^8.1.0",
+    "@storybook/react": "^8.1.0",
+    "@storybook/test-runner": "^0.19.1",
+    "classnames": "^2.5.1",
+    "react": "^18",
+    "resemblejs": "^5.0.0"
+  }
+}
